add landmark example, update yolact example and x86 uint8 (#370)

+++ examples/tm_classification_uint8.c
@@ -83,9 +83,9 @@ int tengine_classify(const char* model_file, const char* image_file, int img_h, 
     }
 
     /* set the input shape to initial the graph, and prerun graph to infer shape */
-    int img_size      = img_h * img_w * 3;
-    int dims[]        = {1, 3, img_h, img_w};    // nchw
-    uint8_t* input_data = (uint8_t*)malloc(img_size);
+    int img_size = img_h * img_w * 3;
+    int dims[] = {1, 3, img_h, img_w};    // nchw
+    uint8_t* input_data = ( uint8_t* )malloc(img_size);
 
     tensor_t input_tensor = get_graph_input_tensor(graph, 0, 0);
     if (input_tensor == NULL)
@@ -139,22 +139,24 @@ int tengine_classify(const char* model_file, const char* image_file, int img_h, 
     }
     fprintf(stderr, "\nmodel file : %s\n", model_file);
     fprintf(stderr, "image file : %s\n", image_file);
-    fprintf(stderr, "img_h, img_w, scale[3], mean[3] : %d %d , %.3f %.3f %.3f, %.1f %.1f %.1f\n", img_h, img_w, scale[0], scale[1], scale[2], mean[0], mean[1], mean[2]);
-    fprintf(stderr, "Repeat %d times, thread %d, avg time %.2f ms, max_time %.2f ms, min_time %.2f ms\n", loop_count, num_thread, total_time / loop_count, max_time, min_time);
+    fprintf(stderr, "img_h, img_w, scale[3], mean[3] : %d %d , %.3f %.3f %.3f, %.1f %.1f %.1f\n", img_h, img_w,
+            scale[0], scale[1], scale[2], mean[0], mean[1], mean[2]);
+    fprintf(stderr, "Repeat %d times, thread %d, avg time %.2f ms, max_time %.2f ms, min_time %.2f ms\n", loop_count,
+            num_thread, total_time / loop_count, max_time, min_time);
     fprintf(stderr, "--------------------------------------\n");
 
     /* get the result of classification */
     tensor_t output_tensor = get_graph_output_tensor(graph, 0, 0);
-    uint8_t* output_u8     = (uint8_t*)get_tensor_buffer(output_tensor);
-    int output_size        = get_tensor_buffer_size(output_tensor);
+    uint8_t* output_u8 = ( uint8_t* )get_tensor_buffer(output_tensor);
+    int output_size = get_tensor_buffer_size(output_tensor);
 
     /* dequant */
     float output_scale = 0.f;
     int output_zero_point = 0;
     get_tensor_quant_param(output_tensor, &output_scale, &output_zero_point, 1);
-    float* output_data = (float*)malloc(output_size*sizeof(float));
-    for (int i=0; i<output_size; i++)
-        output_data[i] = ((float)output_u8[i] - output_zero_point) * output_scale;
+    float* output_data = ( float* )malloc(output_size * sizeof(float));
+    for (int i = 0; i < output_size; i++)
+        output_data[i] = (( float )output_u8[i] - ( float )output_zero_point) * output_scale;
 
     print_topk(output_data, output_size, 5);
     fprintf(stderr, "--------------------------------------\n");
@@ -173,16 +175,20 @@ int tengine_classify(const char* model_file, const char* image_file, int img_h, 
 
 void show_usage()
 {
-    fprintf(stderr, "[Usage]:  [-h]\n    [-m model_file] [-i image_file]\n [-g img_h,img_w] [-s scale[0],scale[1],scale[2]] [-w "
-           "mean[0],mean[1],mean[2]] [-r loop_count] [-t thread_count]\n");
-    fprintf(stderr, "\nmobilenet example: \n    ./classification -m /path/to/mobilenet.tmfile -i /path/to/img.jpg -g 224,224 -s "
-           "0.017,0.017,0.017 -w 104.007,116.669,122.679\n");
+    fprintf(
+        stderr,
+        "[Usage]:  [-h]\n    [-m model_file] [-i image_file]\n [-g img_h,img_w] [-s scale[0],scale[1],scale[2]] [-w "
+        "mean[0],mean[1],mean[2]] [-r loop_count] [-t thread_count]\n");
+    fprintf(
+        stderr,
+        "\nmobilenet example: \n    ./classification -m /path/to/mobilenet.tmfile -i /path/to/img.jpg -g 224,224 -s "
+        "0.017,0.017,0.017 -w 104.007,116.669,122.679\n");
 }
 
 int main(int argc, char* argv[])
 {
-    int loop_count   = DEFAULT_LOOP_COUNT;
-    int num_thread   = DEFAULT_THREAD_COUNT;
+    int loop_count = DEFAULT_LOOP_COUNT;
+    int num_thread = DEFAULT_THREAD_COUNT;
     char* model_file = NULL;
     char* image_file = NULL;
     float img_hw[2] = {0.f};
